fix: restore virtual user username uniqueness and inline DataSourceManager
Keep virtual-user validation independent of multi-data-source username rules,
and merge the thin managers.py into models.py.
